pyproject.toml
[tool.poetry]
name = "gh-action_slack-notify"
version = "0.1.0"
description = "GitHub action to send Slack notifications for failed GitHub Checks."
authors = ["Engineering Experience Squad <platform.eng-ex@sonarsource.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pygithub = "^2.6.1"
slack-sdk = "^3.35.0"
gql = {extras = ["requests"], version = "^3.5.2"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.8"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=src --cov-report html:build/coverage-html --cov-report xml:build/coverage.xml --cov-report term"
pythonpath = "src"
testpaths = "tests"

[tool.autopep8]
max_line_length = 127
